hashes.db
A sqlite-backed index and hashing system for object server contents.

This change removes rsync replication support.  All changes to object server
contents have to go through swift (i.e. ssync), because indexes need to be
updated when the contents change.

The point of all this is that the replicator will no longer need to walk the
filesystem constantly.  Instead, all replication operations will use indexed
database queries.

It maintains compatibility with the terrible hshes.pkl hashing strategy.  For
now.  We can add things later, it's a database!

As a bonus, it also fixes that .ts file hanging around forever thing.  Because
we can just look in the database to see if there are any old .ts files.

Change-Id: Ie4aabef2fd65285648da4037bb1930202b6c8acf